+++ crates/cli/src/commands/mod.rs
@@ -12,6 +12,7 @@ mod list_remote;
 mod migrate;
 mod outdated;
 mod pin;
+mod plugin;
 mod plugins;
 mod remove_plugin;
 mod run;
@@ -36,6 +37,7 @@ pub use list_remote::*;
 pub use migrate::*;
 pub use outdated::*;
 pub use pin::*;
+pub use plugin::*;
 pub use plugins::*;
 pub use remove_plugin::*;
 pub use run::*;
