+++ crates/cli/src/commands/plugin/mod.rs
@@ -1,0 +1,5 @@
+mod add;
+mod remove;
+
+pub use add::*;
+pub use remove::*;
