+++ crates/cli/src/main.rs
@@ -4,7 +4,7 @@ mod error;
 mod helpers;
 mod shell;
 
-use app::{App as CLI, Commands};
+use app::{App as CLI, Commands, PluginCommands};
 use clap::Parser;
 use starbase::{tracing::TracingOptions, App, MainResult};
 use starbase_utils::string_vec;
@@ -48,7 +48,7 @@ async fn main() -> MainResult {
     let mut app = App::new();
 
     match cli.command {
-        Commands::AddPlugin(args) => app.execute_with_args(commands::add_plugin, args),
+        Commands::AddPlugin(args) => app.execute_with_args(commands::add_plugin_old, args),
         Commands::Alias(args) => app.execute_with_args(commands::alias, args),
         Commands::Bin(args) => app.execute_with_args(commands::bin, args),
         Commands::Clean(args) => app.execute_with_args(commands::clean, args),
@@ -61,8 +61,12 @@ async fn main() -> MainResult {
         Commands::Migrate(args) => app.execute_with_args(commands::migrate, args),
         Commands::Outdated(args) => app.execute_with_args(commands::outdated, args),
         Commands::Pin(args) => app.execute_with_args(commands::pin, args),
+        Commands::Plugin { command } => match command {
+            PluginCommands::Add(args) => app.execute_with_args(commands::add_plugin, args),
+            PluginCommands::Remove(args) => app.execute_with_args(commands::remove_plugin, args),
+        },
         Commands::Plugins(args) => app.execute_with_args(commands::plugins, args),
-        Commands::RemovePlugin(args) => app.execute_with_args(commands::remove_plugin, args),
+        Commands::RemovePlugin(args) => app.execute_with_args(commands::remove_plugin_old, args),
         Commands::Run(args) => app.execute_with_args(commands::run, args),
         Commands::Setup(args) => app.execute_with_args(commands::setup, args),
         Commands::Tools(args) => app.execute_with_args(commands::tools, args),
